refactor(flip-game): replace Square class with plain serialisable type
Split domain logic into square.ts and keep squareSlice.ts as pure
Redux wiring. Tests split accordingly.
